Use VSM_Status() to check the status of the Varnish management process.
If the process has changed or restarted, re-initialize the shared
memory attachment and re-acquire the log. A new reader metric
mgt_restart is incremented when the status change is detected.

We run the VSM_Status() check when end-of-log is reached, or when
a flush has been signaled, to minimize the hit from the fstat()
calls in its implementation.

None of this is done when reading from a log file.