Fix declarations and linking of global variables.
The declarations in trackrdrd.h are extern, and the "real" declarations
are in C sources. As it was, it never should have worked; perhaps
ld was forgiving this error in the past, but newer versions do not
allow it.

The same fixes are required for unit test code. This requires
new organization of the object files to link (in Makefile.am).

While here, add an include guard to trackrdrd.h.